+++ tools/tsp-client/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release
 
+## 2024-07-02 - 0.9.1
+
+- Fix error logging after the `compile()` call and exit if diagnostics are encountered.
+- Use `formatDiagnostic()` from "@typespec/compiler" to report diagnostics after compiling.
+
 ## 2024-06-21 - 0.9.0
 
 - Prefer the `service-dir` parameter in the emitter configurations in tspconfig.yaml if specified.